Now test getting the Jinja2 template in tests
