+++ src/itpe/jinja_helpers.py
@@ -1,5 +1,9 @@
 # -*- encoding: utf-8
 
+from jinja2 import Environment, PackageLoader
+
+from .dreamwidth import render_user_links
+
 
 def condense_into_single_line(text):
     """
@@ -8,3 +12,18 @@ def condense_into_single_line(text):
     """
     lines = [line.lstrip() for line in text.split('\n')]
     return ''.join(lines)
+
+
+def get_jinja2_template():
+    """Set up the Jinja2 environment."""
+    env = Environment(
+        loader=PackageLoader('itpe', 'templates'),
+        trim_blocks=True
+    )
+
+    env.filters['condense'] = condense_into_single_line
+    env.filters['userlinks'] = render_user_links
+
+    template = env.get_template('podfic_template.html')
+
+    return template
